+++ mdbx_dump.c
@@ -1,4 +1,4 @@
-/* This file is part of the libmdbx amalgamated source code (v0.14.2-224-g8f756694 at 2026-06-21T11:47:59+03:00).
+/* This file is part of the libmdbx amalgamated source code (v0.14.2-239-gf02137ac at 2026-06-29T13:06:03+03:00).
  *
  * libmdbx (aka MDBX) is an extremely fast, compact, powerful, embeddedable, transactional key-value storage engine with
  * open-source code. MDBX has a specific set of properties and capabilities, focused on creating unique lightweight
@@ -71,7 +71,7 @@ static const flagbit dbflags[] = {{MDBX_REVERSEKEY, "reversekey"},
                                   {0, nullptr}};
 
 static void dumpval(const MDBX_val *v) {
-  static const char digits[] = "0123456789abcdef";
+  const char alpha_offset = 'a' - '9' - 1;
   putchar(' ');
   for (const unsigned char *c = v->iov_base, *end = c + v->iov_len; c < end; ++c) {
     if (mode & PRINT) {
@@ -81,8 +81,10 @@ static void dumpval(const MDBX_val *v) {
       } else
         putchar('\\');
     }
-    putchar(digits[*c >> 4]);
-    putchar(digits[*c & 15]);
+    const int8_t hi = *c >> 4;
+    const int8_t lo = *c & 15;
+    putchar('0' + hi + (((9 - hi) >> 7) & alpha_offset));
+    putchar('0' + lo + (((9 - lo) >> 7) & alpha_offset));
   }
   putchar('\n');
 }